+++ BlueprintUI/Sources/Environment/EnvironmentReader.swift
@@ -29,7 +29,7 @@ public struct EnvironmentReader: Element {
         }
     }
 
-    public func backingViewDescription(bounds: CGRect, subtreeExtent: CGRect?) -> ViewDescription? {
+    public func backingViewDescription(with context: ViewDescriptionContext) -> ViewDescription? {
         return nil
     }
 }