+++ BlueprintUI/Sources/Layout/Empty.swift
@@ -19,7 +19,7 @@ public struct Empty : Element {
         ElementContent(intrinsicSize: .zero)
     }
     
-    public func backingViewDescription(bounds: CGRect, subtreeExtent: CGRect?) -> ViewDescription? {
+    public func backingViewDescription(with context: ViewDescriptionContext) -> ViewDescription? {
         nil
     }
 }