+++ BlueprintUI/Sources/Layout/Overlay.swift
@@ -36,7 +36,7 @@ public struct Overlay: Element {
         }
     }
 
-    public func backingViewDescription(bounds: CGRect, subtreeExtent: CGRect?) -> ViewDescription? {
+    public func backingViewDescription(with context: ViewDescriptionContext) -> ViewDescription? {
         return nil
     }
 }