+++ BlueprintUI/Tests/ConstrainedSizeTests.swift
@@ -331,7 +331,7 @@ fileprivate struct FixedElement: Element {
         }
     }
 
-    func backingViewDescription(bounds: CGRect, subtreeExtent: CGRect?) -> ViewDescription? {
+    func backingViewDescription(with context: ViewDescriptionContext) -> ViewDescription? {
         return nil
     }
 }
@@ -354,7 +354,7 @@ fileprivate struct FlexibleElement: Element {
         }
     }
 
-    func backingViewDescription(bounds: CGRect, subtreeExtent: CGRect?) -> ViewDescription? {
+    func backingViewDescription(with context: ViewDescriptionContext) -> ViewDescription? {
         return nil
     }
 }