+++ BlueprintUI/Tests/GridRowTests.swift
@@ -276,7 +276,7 @@ private struct TestElement: Element {
         return ElementContent(intrinsicSize: size)
     }
 
-    func backingViewDescription(bounds: CGRect, subtreeExtent: CGRect?) -> ViewDescription? {
+    func backingViewDescription(with context: ViewDescriptionContext) -> ViewDescription? {
         return nil
     }
 }